refactored DLR API
