+++ docs/src/lib/dlr.md
@@ -1,0 +1,5 @@
+# Discrete Lehmann Representation
+
+```@autodocs
+Modules = [QuantumStatistics.DLR]
+```
