+++ src/QuantumStatistics.jl
@@ -4,14 +4,27 @@ include("common.jl")
 export σx, σy, σz, σ0
 
 include("grid/grid.jl")
+export Grid
+
 include("fastmath.jl")
+export FastMath
+
 include("utility/utility.jl")
+export Utility
+
 include("correlator/twopoint.jl")
+export TwoPoint
+
 include("correlator/spectral.jl")
+export Spectral
+
 include("correlator/diagram.jl")
-include("correlator/basis.jl")
+export Diagram
+
 include("correlator/dlr/dlr.jl")
+export DLR
+
 include("montecarlo/montecarlo.jl")
-export TwoPoint, Grid, FastMath, Diagram, Spectral, MonteCarlo, Utility, Basis, DLR
+export MonteCarlo
 
 end # module
