--- src/correlator/basis/dlr_fermi/dlr1000000_1e-4.dat
@@ -1,46 +1,0 @@
-           1  -999464.73080965085      -0.44444627912745027          -208649
-           2  -701227.41949596792      -0.20758996658144760           -78223
-           3  -399386.29025201604      -0.13553314923109092           -43726
-           4  -215143.04313868756       -8.7653427436996001E-002      -24369
-           5  -111111.56978186256       -4.5853076105779017E-002      -13273
-           6  -61670.783273361027       -2.8588639180469288E-002       -7133
-           7  -36572.721639061419       -1.6430681706775873E-002       -3890
-           8  -18286.360819530710       -9.5485537772671790E-003       -2095
-           9  -8215.3408533879374       -3.9437787492124406E-003       -1159
-          10  -5478.3392148122493       -2.6157817721648817E-003        -634
-          11  -3573.5798975586608       -1.5601026962969377E-003        -347
-          12  -1621.7966139175592       -9.8594468730311014E-004        -190
-          13  -1097.7344690044056       -5.9678461107919869E-004        -105
-          14  -748.38943584720289       -2.2899412887444930E-004         -78
-          15  -358.22715707639856       -1.4286219390258369E-004         -57
-          16  -240.90149716156651       -8.1743180380152553E-005         -30
-          17  -163.48636076030508       -3.5715548475645922E-005         -16
-          18  -74.598076384899827       -2.3387169870225089E-005          -9
-          19  -34.304202156387674       -1.4312133054653081E-005          -5
-          20  -17.857774237822959       -6.3351430231154663E-006          -4
-          21  -8.9288871189114793       -3.1649509717942597E-006          -3
-          22  -4.4644435594557397       -1.5352433738400875E-006          -2
-          23   2.0418899167981347E-003  -4.0837798335962695E-009          -1
-          24   1.7466255305314498        4.0837798335962695E-009           0
-          25   4.4644435594557406        1.5352433738400875E-006           1
-          26   9.3247595481124801        3.4932510610628993E-006           2
-          27   18.649519096224960        7.1560665273265407E-006           3
-          28   39.027577358718858        1.2188302314819826E-005           4
-          29   74.598076384899841        1.8649519096224959E-005           8
-          30   163.48636076030510        4.6774339740450179E-005          15
-          31   242.96785158711123        7.4598076384899836E-005          29
-          32   358.22715707639856        1.2045074858078326E-004          56
-          33   748.38943584720289        2.9839230553959934E-004         101
-          34   1097.7344690044056        5.0123776135373404E-004         180
-          35   1621.7966139175594        7.8005134814846885E-004         337
-          36   3573.5798975586613        2.1177054567357956E-003         630
-          37   5478.3392148122493        3.3616100490419937E-003        1146
-          38   8215.3408533879374        5.4783392148122501E-003        2073
-          39   18286.360819530713        8.4708218269431822E-003        3885
-          40   36572.721639061427        1.2974372911340475E-002        7258
-          41   61670.783273361027        2.1913356859249000E-002       13409
-          42   111111.56978186258        4.1852508354638107E-002       24158
-          43   215143.04313868759        7.3145443278122849E-002       42965
-          44   399386.29025201604       0.14629088655624570            81877
-          45   701227.41949596792       0.29258177311249139           194287
-          46   999464.73080965085       0.41517993316289520           436523
