--- src/correlator/basis/dlr_fermi/dlr1000000_1e-6.dat
@@ -1,71 +1,0 @@
-           1  -999464.73080965085      -0.44444627912745027          -489708
-           2  -888892.55825490050      -0.35061370974798400          -239292
-           3  -701227.41949596792      -0.22222313956372514          -137560
-           4  -542132.59692436364      -0.16741003341855243           -90472
-           5  -399386.29025201604      -0.12831686690655592           -65583
-           6  -292581.77311249136       -7.6388430218137432E-002      -49622
-           7  -215143.04313868756       -4.1852508354638107E-002      -29769
-           8  -152776.86043627484       -3.1099885003150238E-002      -21521
-           9  -111111.56978186256       -2.2926538052889509E-002      -16035
-          10  -79928.478430656207       -1.2480821570375502E-002      -11749
-          11  -61670.783273361027       -9.5485537772671790E-003       -8600
-          12  -36572.721639061419       -7.3278121239823777E-003       -6322
-          13  -26892.880392335945       -4.5715902048826780E-003       -4627
-          14  -18286.360819530710       -3.4722365556832053E-003       -2518
-          15  -12974.372911340473       -2.7391696074061250E-003       -1846
-          16  -8215.3408533879374       -1.6808050245209968E-003       -1348
-          17  -5478.3392148122493       -1.0024755227074681E-003        -724
-          18  -3573.5798975586608       -7.8005134814846885E-004        -538
-          19  -2387.1384443167944       -5.9678461107919869E-004        -399
-          20  -1621.7966139175592       -3.9002567407423443E-004        -221
-          21  -1097.7344690044056       -2.3784623666581349E-004        -160
-          22  -748.38943584720289       -1.4919615276979967E-004        -117
-          23  -513.45880333674609       -1.1692717322700639E-004         -64
-          24  -358.22715707639856       -9.3548679480900357E-005         -46
-          25  -240.90149716156651       -5.7248532218612326E-005         -34
-          26  -163.48636076030508       -3.9027577358718856E-005         -18
-          27  -111.67437179870815       -2.5340572092461865E-005         -13
-          28  -74.598076384899827       -1.7152101078193837E-005         -10
-          29  -50.681144184923724       -1.1194598658637455E-005          -8
-          30  -34.304202156387674       -8.0227938021366567E-006          -6
-          31  -25.340572092461862       -5.3499406394650879E-006          -5
-          32  -17.857774237822959       -2.7509473614101435E-006          -4
-          33  -8.9288871189114793       -1.2942515081345341E-006          -3
-          34  -4.4644435594557397       -4.7332800392345949E-007          -2
-          35  -2.1440126347742297       -4.0837798335962695E-009          -1
-          36   2.0418899167981347E-003   4.0837798335962695E-009           0
-          37   1.7466255305314498        4.7332800392345949E-007           1
-          38   4.4644435594557406        1.2942515081345341E-006           2
-          39   9.3247595481124801        2.7509473614101435E-006           3
-          40   18.649519096224960        5.3499406394650879E-006           4
-          41   27.126848091275043        8.0227938021366567E-006           5
-          42   39.027577358718858        1.1194598658637455E-005           7
-          43   54.253696182550087        1.7152101078193837E-005           9
-          44   74.598076384899841        2.5340572092461865E-005          12
-          45   105.05031403256230        3.9027577358718856E-005          17
-          46   163.48636076030510        5.5837185899354078E-005          33
-          47   242.96785158711123        7.4598076384899836E-005          45
-          48   358.22715707639856        9.3548679480900357E-005          61
-          49   513.45880333674609        1.4919615276979967E-004          84
-          50   748.38943584720289        2.3385434645401278E-004         115
-          51   1097.7344690044056        3.7419471792360143E-004         158
-          52   1621.7966139175594        5.7144877561033475E-004         221
-          53   2387.1384443167949        7.4838943584720286E-004         397
-          54   3573.5798975586613        9.6360598864626606E-004         540
-          55   5478.3392148122493        1.5601026962969377E-003         737
-          56   8215.3408533879374        2.6157817721648817E-003        1351
-          57   12974.372911340475        3.3616100490419937E-003        1772
-          58   18286.360819530713        4.2354109134715911E-003        2489
-          59   26892.880392335948        6.7232200980839874E-003        3362
-          60   36572.721639061427        1.1463269026444754E-002        4631
-          61   61670.783273361027        1.6039608363319489E-002        8698
-          62   79928.478430656221        2.1913356859249000E-002       11746
-          63   111111.56978186258        3.9964239215328108E-002       16073
-          64   152776.86043627487        5.5555784890931284E-002       22053
-          65   215143.04313868759        7.6388430218137432E-002       29406
-          66   292581.77311249141       0.11724499398371804            50232
-          67   399386.29025201604       0.15277686043627486            68239
-          68   542132.59692436364       0.19158769557688393            96107
-          69   701227.41949596792       0.29258177311249139           139338
-          70   888892.55825490062       0.36682460884623214           213075
-          71   999464.73080965085       0.46897997593487217           470885
