--- src/correlator/basis/dlr_fermi/dlr1000000_1e-8.dat
@@ -1,92 +1,0 @@
-           1  -999464.73080965085      -0.44444627912745027          -922767
-           2  -888892.55825490050      -0.35061370974798400          -455940
-           3  -798772.58050403208      -0.28102002406512783          -208221
-           4  -701227.41949596792      -0.21514304313868760          -136165
-           5  -542132.59692436364      -0.15985695686131243           -94296
-           6  -399386.29025201604      -0.12177727317289651           -66676
-           7  -292581.77311249136       -9.1706152211558034E-002      -49015
-           8  -215143.04313868756       -6.7766574615545458E-002      -36989
-           9  -183412.30442311606       -4.9923286281502006E-002      -28408
-          10  -152776.86043627484       -3.6572721639061424E-002      -21685
-          11  -111111.56978186256       -3.1550229993699525E-002      -15700
-          12  -79928.478430656207       -2.6892880392335949E-002      -13408
-          13  -61670.783273361027       -1.9982119607664054E-002      -11507
-          14  -45853.076105779015       -1.4294319590234644E-002       -8537
-          15  -36572.721639061419       -1.0463127088659527E-002       -6301
-          16  -26892.880392335945       -7.8875574984248811E-003       -4618
-          17  -21913.356859248997       -5.9871154867776229E-003       -3415
-          18  -18286.360819530710       -4.5715902048826780E-003       -2543
-          19  -12974.372911340473       -3.4722365556832053E-003       -1857
-          20  -10463.127088659525       -2.7391696074061250E-003       -1586
-          21  -8215.3408533879374       -2.1954689380088111E-003       -1357
-          22  -5478.3392148122493       -1.4329086283055943E-003        -994
-          23  -4390.9378760176223       -1.1428975512206695E-003        -735
-          24  -3573.5798975586608       -8.9339497438966525E-004        -536
-          25  -2387.1384443167944       -5.7144877561033475E-004        -392
-          26  -1621.7966139175592       -4.5798825774889861E-004        -286
-          27  -1307.8908860824406       -3.5822715707639857E-004        -212
-          28  -1097.7344690044056       -2.7443361725110139E-004        -157
-          29  -748.38943584720289       -2.2334874359741631E-004        -116
-          30  -624.44123773950162       -1.4919615276979967E-004         -89
-          31  -513.45880333674609       -1.2220099345467507E-004         -65
-          32  -358.22715707639856       -9.7506418518558606E-005         -47
-          33  -298.39230553959931       -7.8055154717437711E-005         -40
-          34  -240.90149716156651       -6.2654720169216756E-005         -34
-          35  -163.48636076030508       -4.0871590190076277E-005         -24
-          36  -111.67437179870815       -3.2091175208546627E-005         -18
-          37  -89.556789269099639       -2.5340572092461865E-005         -13
-          38  -74.598076384899827       -1.6544573880748403E-005         -10
-          39  -50.681144184923724       -1.1693584935112545E-005          -8
-          40  -34.304202156387674       -8.2722869403742014E-006          -7
-          41  -25.340572092461862       -5.5972993293187277E-006          -6
-          42  -17.857774237822959       -3.3413692617015407E-006          -5
-          43  -8.9288871189114793       -2.2794538917849129E-006          -4
-          44  -4.4644435594557397       -1.5352433738400875E-006          -3
-          45  -2.1440126347742297       -4.7332800392345949E-007          -2
-          46   2.0418899167981347E-003  -4.0837798335962695E-009          -1
-          47   1.7466255305314498        4.0837798335962695E-009           0
-          48   4.4644435594557406        4.7332800392345949E-007           1
-          49   9.3247595481124801        1.5352433738400875E-006           2
-          50   14.312133054653081        3.1649509717942597E-006           3
-          51   14.615896653375799        4.4644435594557403E-006           4
-          52   18.649519096224960        6.0941511574099129E-006           5
-          53   27.126848091275043        9.3247595481124795E-006           6
-          54   39.027577358718858        1.3131289254070288E-005           7
-          55   54.253696182550087        1.9513788679359428E-005           9
-          56   74.598076384899841        2.9730779583226687E-005          13
-          57   105.05031403256230        3.7299038192449918E-005          18
-          58   132.35659104598722        4.6774339740450179E-005          21
-          59   163.48636076030510        6.8608404312775348E-005          25
-          60   242.96785158711123        8.5599050231441407E-005          34
-          61   358.22715707639856        1.1167437179870816E-004          47
-          62   434.02956946040069        1.4286219390258369E-004          64
-          63   513.45880333674609        1.7911357853819929E-004          87
-          64   748.38943584720289        2.2899412887444930E-004         118
-          65   1097.7344690044056        2.9839230553959934E-004         162
-          66   1369.5848037030623        4.4669748719483262E-004         190
-          67   1621.7966139175594        5.7144877561033475E-004         224
-          68   2387.1384443167949        7.1645431415279714E-004         301
-          69   3573.5798975586613        1.0977344690044056E-003         409
-          70   4390.9378760176223        1.4329086283055943E-003         555
-          71   5478.3392148122493        1.8319530309955944E-003         750
-          72   8215.3408533879374        2.2857951024413390E-003        1018
-          73   10463.127088659527        2.9935577433888114E-003        1378
-          74   12974.372911340475        3.8055397866530159E-003        1866
-          75   18286.360819530713        4.7742768886335895E-003        2536
-          76   21913.356859248997        7.6110795733060317E-003        2946
-          77   26892.880392335948        9.9910598038320270E-003        3423
-          78   36572.721639061427        1.2974372911340475E-002        4612
-          79   45853.076105779015        1.7563751504070489E-002        6240
-          80   61670.783273361027        2.2926538052889509E-002        8409
-          81   79928.478430656221        3.0835391636680514E-002       11516
-          82   111111.56978186258        4.1852508354638107E-002       15709
-          83   152776.86043627487        5.8622496991859022E-002       20944
-          84   183412.30442311606        7.6388430218137432E-002       29099
-          85   215143.04313868759        9.1706152211558034E-002       36959
-          86   292581.77311249141       0.11724499398371804            48865
-          87   399386.29025201604       0.14629088655624570            64293
-          88   542132.59692436364       0.18341230442311607            90048
-          89   701227.41949596792       0.23448998796743609           136731
-          90   798772.58050403208       0.28102002406512783           174570
-          91   888892.55825490062       0.35061370974798400           236758
-          92   999464.73080965085       0.45741822688750861           491804
