--- src/correlator/basis/dlr_fermi/dlr100000_1e-10.dat
@@ -1,92 +1,0 @@
-           1  -99946.473080965094      -0.46897997593487217           -46312
-           2  -93795.995186974425      -0.38317539115376786           -22300
-           3  -83035.986632579035      -0.31971391372262487           -16421
-           4  -76635.078230753570      -0.25663373381311183           -12777
-           5  -66964.013367420965      -0.20758996658144760            -9955
-           6  -50480.367989919236      -0.16741003341855243            -8251
-           7  -43028.608627737514      -0.13553314923109092            -5964
-           8  -36682.460884623215      -0.11111156978186257            -4532
-           9  -27106.629846218184       -9.1706152211558034E-002       -3416
-          10  -22870.911344375432       -7.3145443278122849E-002       -2652
-          11  -19158.769557688393       -6.2199770006300475E-002       -2270
-          12  -14629.088655624570       -5.3785760784671899E-002       -1925
-          13  -11111.156978186256       -4.1852508354638107E-002       -1440
-          14  -9579.3847788441963       -3.1283454324396817E-002       -1234
-          15  -8370.5016709276206       -2.4961643140751003E-002       -1058
-          16  -6167.0783273361021       -1.9097107554534358E-002        -775
-          17  -5189.7491645361897       -1.4655624247964755E-002        -574
-          18  -4185.2508354638103       -1.1463269026444754E-002        -493
-          19  -3512.7503008140980       -9.1431804097653561E-003        -424
-          20  -2931.1248495929508       -7.3278121239823777E-003        -313
-          21  -2092.6254177319051       -5.9871154867776229E-003        -234
-          22  -1756.3751504070490       -4.7742768886335895E-003        -203
-          23  -1465.5624247964754       -4.0099020908298724E-003        -169
-          24  -1095.6678429624499       -3.2435932278351187E-003        -144
-          25  -914.31804097653560       -2.7391696074061250E-003        -123
-          26  -780.83182094503979       -2.1954689380088111E-003         -90
-          27  -573.16345132223773       -1.7867899487793305E-003         -67
-          28  -477.42768886335892       -1.4329086283055943E-003         -57
-          29  -400.99020908298712       -1.0977344690044056E-003         -49
-          30  -286.58172566111887       -9.1597651549779721E-004         -36
-          31  -238.71384443167946       -7.8005134814846885E-004         -31
-          32  -194.37428126968899       -6.5394544304122042E-004         -26
-          33  -143.29086283055943       -4.5798825774889861E-004         -19
-          34  -109.77344690044056       -3.9002567407423443E-004         -16
-          35  -93.541738581605117       -3.2697272152061021E-004         -14
-          36  -81.089830695877964       -2.6471318209197444E-004         -12
-          37  -59.678461107919865       -2.2334874359741631E-004         -10
-          38  -43.402956946040064       -1.6348636076030511E-004          -8
-          39  -34.239620092576558       -1.2045074858078326E-004          -7
-          40  -24.648617182577752       -8.9556789269099643E-005          -6
-          41  -18.709735896180071       -6.6178295522993611E-005          -5
-          42  -13.235659104598723       -3.5715548475645922E-005          -4
-          43  -7.8055154717437700       -2.2007578891281148E-005          -3
-          44  -4.2799525115720698       -1.0354012065076273E-005          -2
-          45  -1.9513788679359425       -3.7866240313876759E-006          -1
-          46   1.6335119334385077E-003  -3.2670238668770156E-008           0
-          47   1.7857774237822961        3.2670238668770156E-008           1
-          48   4.2799525115720698        3.7866240313876759E-006           2
-          49   8.5599050231441396        1.0354012065076273E-005           3
-          50   13.721680862555070        2.2007578891281148E-005           4
-          51   18.709735896180071        3.5715548475645922E-005           5
-          52   24.090149716156652        4.8753209259279303E-005           6
-          53   31.222061886975084        6.6178295522993611E-005           7
-          54   43.402956946040064        8.9556789269099643E-005           9
-          55   59.678461107919873        1.1892311833290675E-004          11
-          56   78.005134814846883        1.5611030943487542E-004          13
-          57   91.597651549779727        2.1701478473020033E-004          18
-          58   105.88527283678978        2.9839230553959934E-004          21
-          59   143.29086283055943        3.5822715707639857E-004          25
-          60   168.08050245209967        4.3402956946040066E-004          34
-          61   200.49510454149359        5.9678461107919869E-004          40
-          62   238.71384443167949        7.1645431415279714E-004          47
-          63   286.58172566111887        8.9339497438966525E-004          64
-          64   400.99020908298718        1.1428975512206695E-003          75
-          65   477.42768886335898        1.4329086283055943E-003          88
-          66   573.16345132223773        1.7361182778416026E-003         121
-          67   801.98041816597436        2.1177054567357956E-003         142
-          68   954.85537772671796        2.6157817721648817E-003         167
-          69   1146.3269026444755        3.2435932278351187E-003         234
-          70   1541.7695818340258        3.9437787492124406E-003         271
-          71   1828.6360819530712        4.7742768886335895E-003         313
-          72   2092.6254177319051        5.7316345132223771E-003         426
-          73   2858.8639180469290        7.1471597951173220E-003         496
-          74   3512.7503008140980        8.7818757520352446E-003         580
-          75   4185.2508354638103        1.1463269026444754E-002         791
-          76   5862.2496991859025        1.4655624247964755E-002        1063
-          77   7025.5006016281959        1.9097107554534358E-002        1244
-          78   8370.5016709276206        2.1913356859249000E-002        1449
-          79   11111.156978186256        2.5948745822680950E-002        1943
-          80   12831.686690655590        3.3883287307772729E-002        2263
-          81   14629.088655624570        4.5853076105779017E-002        2628
-          82   19158.769557688393        5.9866712692227271E-002        3439
-          83   22870.911344375432        7.0255006016281957E-002        4582
-          84   27106.629846218184        8.3705016709276214E-002        5337
-          85   36682.460884623215       0.10757152156934380             6353
-          86   43028.608627737514       0.14051001203256391             8721
-          87   50480.367989919236       0.17530685487399200            10542
-          88   66964.013367420965       0.21514304313868760            12875
-          89   76635.078230753570       0.26289090730841397            19783
-          90   83035.986632579035       0.31971391372262487            27405
-          91   93795.995186974440       0.38317539115376786            43952
-          92   99946.473080965079       0.45741822688750861            89487
