--- src/correlator/basis/dlr_fermi/dlr100000_1e-4.dat
@@ -1,39 +1,0 @@
-           1  -99946.473080965094      -0.49973236540482546           -21626
-           2  -83035.986632579035      -0.29258177311249139           -12800
-           3  -50480.367989919236      -0.16741003341855243            -8140
-           4  -27106.629846218184      -0.10757152156934380            -4361
-           5  -14629.088655624570       -6.1670783273361028E-002       -2468
-           6  -8370.5016709276206       -3.9964239215328108E-002       -1355
-           7  -4185.2508354638103       -2.4961643140751003E-002        -734
-           8  -2931.1248495929508       -1.1463269026444754E-002        -415
-           9  -2092.6254177319051       -4.9955299019160135E-003        -227
-          10  -1095.6678429624499       -3.2435932278351187E-003        -122
-          11  -573.16345132223773       -2.1954689380088111E-003         -64
-          12  -286.58172566111887       -9.7187140634844493E-004         -34
-          13  -143.29086283055943       -6.5394544304122042E-004         -18
-          14  -81.089830695877964       -4.0544915347938984E-004         -10
-          15  -43.402956946040064       -1.9501283703711721E-004          -6
-          16  -24.648617182577752       -9.3548679480900357E-005          -4
-          17  -13.235659104598723       -4.6774339740450179E-005          -3
-          18  -7.8055154717437700       -2.5319607774354078E-005          -2
-          19  -4.2799525115720698       -1.2281946990720700E-005          -1
-          20   1.6335119334385077E-003  -3.2670238668770156E-008           0
-          21   1.7857774237822961        3.2670238668770156E-008           1
-          22   4.2799525115720698        1.0354012065076273E-005           2
-          23   13.721680862555070        3.5715548475645922E-005           3
-          24   24.090149716156652        7.1431096951291844E-005           4
-          25   43.402956946040064        1.3235659104598722E-004           8
-          26   78.005134814846883        2.6471318209197444E-004          15
-          27   143.29086283055943        3.9002567407423443E-004          29
-          28   286.58172566111887        5.7144877561033475E-004          41
-          29   573.16345132223773        1.3695848037030625E-003          57
-          30   801.98041816597436        3.1202053925938754E-003         109
-          31   1146.3269026444755        4.9955299019160135E-003         208
-          32   2092.6254177319051        7.8875574984248811E-003         389
-          33   4185.2508354638103        1.1974230973555246E-002         732
-          34   8370.5016709276206        1.8286360819530712E-002        1356
-          35   14629.088655624570        2.9933356346113636E-002        2458
-          36   27106.629846218184        4.9923286281502006E-002        4409
-          37   50480.367989919236        8.3705016709276214E-002        8440
-          38   83035.986632579035       0.15985695686131243            19669
-          39   99946.473080965079       0.29258177311249139            44003
