--- src/correlator/basis/dlr_fermi/dlr100000_1e-6.dat
@@ -1,59 +1,0 @@
-           1  -99946.473080965094      -0.48710909269158603           -46544
-           2  -83035.986632579035      -0.41517993316289520           -22331
-           3  -66964.013367420965      -0.29258177311249139           -12816
-           4  -50480.367989919236      -0.20758996658144760            -8261
-           5  -36682.460884623215      -0.13144545365420698            -5968
-           6  -27106.629846218184       -8.3705016709276214E-002       -4562
-           7  -19158.769557688393       -6.5722726827103492E-002       -2655
-           8  -14629.088655624570       -5.1897491645361900E-002       -1930
-           9  -8370.5016709276206       -3.2861363413551746E-002       -1438
-          10  -6167.0783273361021       -1.9982119607664054E-002        -776
-          11  -4185.2508354638103       -1.5775114996849762E-002        -573
-          12  -2931.1248495929508       -1.2480821570375502E-002        -424
-          13  -2092.6254177319051       -7.7088479091701285E-003        -235
-          14  -1465.5624247964754       -5.2315635443297634E-003        -170
-          15  -1095.6678429624499       -3.4722365556832053E-003        -124
-          16  -780.83182094503979       -2.4977649509580068E-003         -91
-          17  -573.16345132223773       -1.7361182778416026E-003         -68
-          18  -400.99020908298712       -1.1935692221583974E-003         -49
-          19  -286.58172566111887       -8.4040251226049842E-004         -36
-          20  -194.37428126968899       -5.7144877561033475E-004         -19
-          21  -143.29086283055943       -4.0544915347938984E-004         -14
-          22  -81.089830695877964       -2.8572438780516738E-004         -10
-          23  -59.678461107919865       -2.0272457673969492E-004          -8
-          24  -43.402956946040064       -1.3721680862555070E-004          -6
-          25  -24.648617182577752       -7.4598076384899836E-005          -5
-          26  -18.709735896180071       -4.0871590190076277E-005          -4
-          27  -13.235659104598723       -2.2007578891281148E-005          -3
-          28  -7.8055154717437700       -1.0354012065076273E-005          -2
-          29  -4.2799525115720698       -3.7866240313876759E-006          -1
-          30   1.6335119334385077E-003  -3.2670238668770156E-008           0
-          31   1.7857774237822961        3.2670238668770156E-008           1
-          32   4.2799525115720698        3.7866240313876759E-006           2
-          33   8.5599050231441396        1.0354012065076273E-005           3
-          34   13.721680862555070        2.2007578891281148E-005           4
-          35   24.090149716156652        4.0871590190076277E-005           5
-          36   43.402956946040064        7.4598076384899836E-005           9
-          37   59.678461107919873        1.4286219390258369E-004          13
-          38   78.005134814846883        1.9501283703711721E-004          18
-          39   143.29086283055943        2.7443361725110139E-004          25
-          40   200.49510454149359        3.9002567407423443E-004          35
-          41   286.58172566111887        5.9678461107919869E-004          48
-          42   400.99020908298718        8.9339497438966525E-004          65
-          43   573.16345132223773        1.3078908860824408E-003          91
-          44   801.98041816597436        1.9718893746062203E-003         122
-          45   1146.3269026444755        2.8658172566111886E-003         169
-          46   1541.7695818340258        4.3909378760176223E-003         236
-          47   2092.6254177319051        5.4783392148122501E-003         320
-          48   2858.8639180469290        6.9444731113664105E-003         420
-          49   4185.2508354638103        1.0463127088659527E-002         758
-          50   5862.2496991859025        1.5641727162198409E-002        1027
-          51   8370.5016709276206        2.4961643140751003E-002        1432
-          52   14629.088655624570        3.2861363413551746E-002        1996
-          53   19158.769557688393        4.5853076105779017E-002        2669
-          54   27106.629846218184        6.0888636586448254E-002        4641
-          55   36682.460884623215        7.9928478430656216E-002        6288
-          56   50480.367989919236       0.15277686043627486             8772
-          57   66964.013367420965       0.21514304313868760            12986
-          58   83035.986632579035       0.30555372087254973            19980
-          59   99946.473080965079       0.43028608627737519            44646
