--- src/correlator/basis/dlr_fermi/dlr100000_1e-8.dat
@@ -1,75 +1,0 @@
-           1  -99946.473080965094      -0.41517993316289520           -43834
-           2  -93795.995186974425      -0.35061370974798400           -28072
-           3  -83035.986632579035      -0.28102002406512783           -19809
-           4  -66964.013367420965      -0.21514304313868760           -12823
-           5  -50480.367989919236      -0.17530685487399200            -8680
-           6  -43028.608627737514      -0.11724499398371804            -6265
-           7  -36682.460884623215       -9.1706152211558034E-002       -4513
-           8  -27106.629846218184       -6.7766574615545458E-002       -3386
-           9  -19158.769557688393       -5.1897491645361900E-002       -2579
-          10  -14629.088655624570       -3.9964239215328108E-002       -1898
-          11  -11111.156978186256       -3.0444318293224127E-002       -1409
-          12  -8370.5016709276206       -2.1913356859249000E-002       -1033
-          13  -6167.0783273361021       -1.6941643653886364E-002        -762
-          14  -4185.2508354638103       -1.3446440196167975E-002        -560
-          15  -3512.7503008140980       -9.5485537772671790E-003        -410
-          16  -2931.1248495929508       -7.8208635810992043E-003        -304
-          17  -2092.6254177319051       -6.2404107851877508E-003        -229
-          18  -1465.5624247964754       -4.9955299019160135E-003        -164
-          19  -1095.6678429624499       -4.1076704266939683E-003        -139
-          20  -780.83182094503979       -2.7391696074061250E-003        -119
-          21  -573.16345132223773       -1.8708347716321022E-003         -87
-          22  -477.42768886335892       -1.4967788716944057E-003         -64
-          23  -400.99020908298712       -1.2488824754790034E-003         -46
-          24  -286.58172566111887       -8.4040251226049842E-004         -34
-          25  -194.37428126968899       -6.8479240185153126E-004         -25
-          26  -143.29086283055943       -5.4886723450220279E-004         -21
-          27  -109.77344690044056       -3.7419471792360143E-004         -18
-          28  -81.089830695877964       -2.6471318209197444E-004         -13
-          29  -59.678461107919865       -1.8709735896180071E-004         -10
-          30  -43.402956946040064       -1.3721680862555070E-004          -8
-          31  -34.239620092576558       -1.0136228836984746E-004          -7
-          32  -24.648617182577752       -7.4598076384899836E-005          -6
-          33  -18.709735896180071       -3.9027577358718856E-005          -5
-          34  -13.235659104598723       -2.2007578891281148E-005          -4
-          35  -7.8055154717437700       -1.0354012065076273E-005          -3
-          36  -4.2799525115720698       -3.7866240313876759E-006          -2
-          37  -1.9513788679359425       -3.2670238668770156E-008          -1
-          38   1.6335119334385077E-003   3.2670238668770156E-008           0
-          39   1.7857774237822961        3.7866240313876759E-006           1
-          40   4.2799525115720698        1.0354012065076273E-005           2
-          41   8.5599050231441396        2.0163566059923730E-005           3
-          42   13.721680862555070        3.7299038192449918E-005           4
-          43   24.090149716156652        6.8608404312775348E-005           5
-          44   31.222061886975084        9.3548679480900357E-005           6
-          45   43.402956946040064        1.2836470083418651E-004           7
-          46   59.678461107919873        1.7119810046288281E-004           9
-          47   78.005134814846883        2.3784623666581349E-004          13
-          48   105.88527283678978        3.4239620092576563E-004          18
-          49   143.29086283055943        4.9297234365155507E-004          21
-          50   200.49510454149359        5.9678461107919869E-004          25
-          51   238.71384443167949        7.1645431415279714E-004          34
-          52   286.58172566111887        1.0024755227074681E-003          47
-          53   400.99020908298718        1.5601026962969377E-003          64
-          54   477.42768886335898        1.9272119772925321E-003          87
-          55   573.16345132223773        2.3871384443167947E-003         119
-          56   801.98041816597436        3.4722365556832053E-003         163
-          57   1146.3269026444755        5.2315635443297634E-003         193
-          58   1541.7695818340258        6.4871864556702375E-003         228
-          59   2092.6254177319051        8.0198041816597447E-003         305
-          60   2858.8639180469290        1.2480821570375502E-002         411
-          61   3512.7503008140980        1.6039608363319489E-002         556
-          62   4185.2508354638103        2.0926254177319054E-002         746
-          63   5862.2496991859025        2.5948745822680950E-002        1028
-          64   8370.5016709276206        3.2861363413551746E-002        1400
-          65   11111.156978186256        4.1852508354638107E-002        1880
-          66   14629.088655624570        5.3785760784671899E-002        2600
-          67   19158.769557688393        7.0255006016281957E-002        3347
-          68   27106.629846218184        9.5793847788441966E-002        4447
-          69   36682.460884623215       0.13553314923109092             5885
-          70   43028.608627737514       0.15985695686131243             8155
-          71   50480.367989919236       0.20758996658144760            12678
-          72   66964.013367420965       0.25240183994959620            16587
-          73   83035.986632579035       0.30555372087254973            22331
-          74   93795.995186974440       0.38317539115376786            46310
-          75   99946.473080965079       0.44444627912745027            95738
