--- src/correlator/basis/dlr_fermi/dlr10000_1e-10.dat
@@ -1,73 +1,0 @@
-           1  -9994.6473080965079      -0.48710909269158603            -8059
-           2  -9742.1818538317202      -0.43028608627737519            -4286
-           3  -9148.3645377501725      -0.33482006683710486            -2106
-           4  -7663.5078230753579      -0.29258177311249139            -1539
-           5  -7012.2741949596784      -0.24355454634579302            -1191
-           6  -6111.0744174509946      -0.20758996658144760             -921
-           7  -5005.3526919034912      -0.17530685487399200             -758
-           8  -4151.7993316289521      -0.13144545365420698             -544
-           9  -3348.2006683710479      -0.11435455672187715             -472
-          10  -2810.2002406512784       -9.9846572563004013E-002        -411
-          11  -2222.2313956372514       -7.9928478430656216E-002        -308
-          12  -1674.1003341855239       -6.3100459987399049E-002        -239
-          13  -1405.1001203256392       -4.7896923894220983E-002        -201
-          14  -1172.4499398371804       -3.6572721639061424E-002        -169
-          15  -876.53427436995980       -2.9311248495929511E-002        -125
-          16  -763.88430218137432       -2.3948461947110491E-002        -106
-          17  -641.58433453277939       -1.9982119607664054E-002         -91
-          18  -478.96923894220987       -1.6430681706775873E-002         -66
-          19  -418.52508354638098       -1.3888946222732821E-002         -48
-          20  -351.27503008140980       -1.1463269026444754E-002         -41
-          21  -268.92880392335945       -8.2153408533879365E-003         -35
-          22  -199.82119607664052       -5.9871154867776229E-003         -26
-          23  -160.39608363319485       -4.9955299019160135E-003         -22
-          24  -129.74372911340475       -4.2354109134715911E-003         -19
-          25  -104.63127088659525       -3.4722365556832053E-003         -14
-          26  -82.153408533879372       -2.9935577433888114E-003         -12
-          27  -64.871864556702377       -2.1177054567357956E-003         -10
-          28  -49.955299019160130       -1.5601026962969377E-003          -9
-          29  -38.874856253937800       -1.2488824754790034E-003          -8
-          30  -29.935577433888117       -1.0024755227074681E-003          -7
-          31  -23.871384443167948       -6.2444123773950169E-004          -6
-          32  -18.708347716321022       -3.9002567407423443E-004          -5
-          33  -11.935692221583974       -2.3766236932313303E-004          -4
-          34  -7.8005134814846881       -1.4588504907423443E-004          -3
-          35  -3.9002567407423441       -6.8079993869750845E-005          -2
-          36  -1.7119810046288277       -2.0572557091974445E-005          -1
-          37  -3.2391278384334754E-002  -2.6136190935016125E-007           0
-          38   1.7911357853819927        2.6136190935016125E-007           1
-          39   3.7419471792360146        2.0572557091974445E-005           2
-          40   6.8479240185153119        6.8079993869750845E-005           3
-          41   11.428975512206694        1.4588504907423443E-004           4
-          42   18.708347716321022        2.3766236932313303E-004           5
-          43   23.871384443167948        3.9002567407423443E-004           6
-          44   29.935577433888117        6.2444123773950169E-004           7
-          45   40.099020908298719        1.0024755227074681E-003           9
-          46   52.315635443297630        1.2488824754790034E-003          11
-          47   64.871864556702377        1.5601026962969377E-003          13
-          48   80.198041816597438        2.1177054567357956E-003          15
-          49   99.910598038320273        2.9935577433888114E-003          18
-          50   129.74372911340475        4.2354109134715911E-003          25
-          51   157.75114996849763        4.9955299019160135E-003          30
-          52   190.97107554534358        5.9871154867776229E-003          35
-          53   249.61643140751005        8.2153408533879365E-003          48
-          54   328.61363413551749        9.5485537772671790E-003          55
-          55   381.94215109068716        1.1463269026444754E-002          66
-          56   458.53076105779013        1.4294319590234644E-002          78
-          57   616.70783273361030        1.7563751504070489E-002          91
-          58   731.45443278122843        2.0926254177319054E-002         125
-          59   837.05016709276208        2.5948745822680950E-002         170
-          60   1111.1156978186257        3.1283454324396817E-002         200
-          61   1527.7686043627486        3.8194215109068716E-002         236
-          62   1753.0685487399198        4.5853076105779017E-002         307
-          63   2151.4304313868761        5.7177278360938576E-002         357
-          64   2566.3373381311180        7.0255006016281957E-002         414
-          65   3197.1391372262487        8.7653427436996001E-002         548
-          66   3831.7539115376790       0.11435455672187715              656
-          67   4933.6626618688824       0.14051001203256391              791
-          68   6111.0744174509946       0.18341230442311607             1132
-          69   7012.2741949596793       0.21514304313868760             1428
-          70   7663.5078230753579       0.25240183994959620             1862
-          71   9148.3645377501725       0.29258177311249139             2608
-          72   9742.1818538317202       0.35061370974798400             4154
-          73   9994.6473080965079       0.45741822688750861             8434
