--- src/correlator/basis/dlr_fermi/dlr10000_1e-12.dat
@@ -1,85 +1,0 @@
-           1  -9994.6473080965079      -0.46897997593487217            -9532
-           2  -9742.1818538317202      -0.41517993316289520            -4538
-           3  -9148.3645377501725      -0.35061370974798400            -2961
-           4  -8605.7217255475025      -0.29258177311249139            -2159
-           5  -7663.5078230753579      -0.24355454634579302            -1619
-           6  -7012.2741949596784      -0.20758996658144760            -1253
-           7  -6111.0744174509946      -0.17530685487399200             -998
-           8  -5005.3526919034912      -0.14629088655624570             -795
-           9  -4151.7993316289521      -0.12620091997479810             -655
-          10  -3668.2460884623210      -0.10757152156934380             -560
-          11  -3348.2006683710479       -9.5793847788441966E-002        -419
-          12  -2810.2002406512784       -7.6388430218137432E-002        -360
-          13  -2222.2313956372514       -6.1670783273361028E-002        -312
-          14  -1915.8769557688395       -4.9923286281502006E-002        -240
-          15  -1674.1003341855239       -4.3826713718498000E-002        -202
-          16  -1405.1001203256392       -3.8194215109068716E-002        -173
-          17  -1172.4499398371804       -3.0835391636680514E-002        -149
-          18  -876.53427436995980       -2.4961643140751003E-002        -126
-          19  -763.88430218137432       -2.0926254177319054E-002        -109
-          20  -641.58433453277939       -1.7563751504070489E-002         -91
-          21  -555.55784890931284       -1.4655624247964755E-002         -78
-          22  -478.96923894220987       -1.1974230973555246E-002         -66
-          23  -418.52508354638098       -9.9910598038320270E-003         -49
-          24  -351.27503008140980       -8.4708218269431822E-003         -42
-          25  -268.92880392335945       -7.1471597951173220E-003         -36
-          26  -229.26538052889507       -6.2404107851877508E-003         -31
-          27  -199.82119607664052       -5.2315635443297634E-003         -26
-          28  -160.39608363319485       -4.3909378760176223E-003         -22
-          29  -129.74372911340475       -3.8055397866530159E-003         -19
-          30  -104.63127088659525       -3.2435932278351187E-003         -16
-          31  -82.153408533879372       -2.7391696074061250E-003         -14
-          32  -64.871864556702377       -2.0049510454149362E-003         -12
-          33  -57.316345132223766       -1.6808050245209968E-003         -10
-          34  -49.955299019160130       -1.3695848037030625E-003          -9
-          35  -38.874856253937800       -9.1597651549779721E-004          -8
-          36  -29.935577433888117       -7.1645431415279714E-004          -7
-          37  -23.871384443167948       -5.7144877561033475E-004          -6
-          38  -18.708347716321022       -3.5822715707639857E-004          -5
-          39  -11.935692221583974       -2.2356806790802556E-004          -4
-          40  -7.8005134814846881       -1.3005409292360146E-004          -3
-          41  -3.9002567407423441       -6.8079993869750845E-005          -2
-          42  -1.7119810046288277       -2.0572557091974445E-005          -1
-          43  -3.2391278384334754E-002  -2.6136190935016125E-007           0
-          44   1.7911357853819927        2.6136190935016125E-007           1
-          45   3.7419471792360146        2.0572557091974445E-005           2
-          46   6.8479240185153119        6.8079993869750845E-005           3
-          47   11.428975512206694        1.3005409292360146E-004           4
-          48   18.708347716321022        2.2356806790802556E-004           5
-          49   23.871384443167948        3.5822715707639857E-004           6
-          50   29.935577433888117        5.7144877561033475E-004           7
-          51   40.099020908298719        7.1645431415279714E-004           8
-          52   45.715902048826777        9.1597651549779721E-004           9
-          53   52.315635443297630        1.3695848037030625E-003          11
-          54   64.871864556702377        1.6808050245209968E-003          13
-          55   80.198041816597438        2.0049510454149362E-003          15
-          56   99.910598038320273        2.7391696074061250E-003          18
-          57   129.74372911340475        3.2435932278351187E-003          21
-          58   157.75114996849763        3.8055397866530159E-003          25
-          59   190.97107554534358        5.2315635443297634E-003          29
-          60   219.13356859248998        5.9871154867776229E-003          34
-          61   249.61643140751005        7.1471597951173220E-003          40
-          62   328.61363413551749        8.4708218269431822E-003          48
-          63   381.94215109068716        9.9910598038320270E-003          57
-          64   458.53076105779013        1.2480821570375502E-002          67
-          65   537.85760784671902        1.4966678173056818E-002          91
-          66   616.70783273361030        1.7563751504070489E-002         107
-          67   731.45443278122843        2.0926254177319054E-002         126
-          68   837.05016709276208        2.5948745822680950E-002         147
-          69   1111.1156978186257        2.8588639180469288E-002         171
-          70   1283.1686690655590        3.2079216726638979E-002         203
-          71   1527.7686043627486        3.9964239215328108E-002         239
-          72   1753.0685487399198        4.5853076105779017E-002         311
-          73   2151.4304313868761        5.3785760784671899E-002         357
-          74   2566.3373381311180        6.7766574615545458E-002         416
-          75   3197.1391372262487        8.7653427436996001E-002         490
-          76   3831.7539115376790        9.9846572563004013E-002         583
-          77   4444.4627912745027       0.11724499398371804              693
-          78   4933.6626618688824       0.14629088655624570              809
-          79   6111.0744174509946       0.17530685487399200              984
-          80   7012.2741949596793       0.21514304313868760             1206
-          81   7663.5078230753579       0.25663373381311183             1492
-          82   8605.7217255475043       0.30555372087254973             1869
-          83   9148.3645377501725       0.35061370974798400             2553
-          84   9742.1818538317202       0.41517993316289520             4154
-          85   9994.6473080965079       0.46897997593487217             8422
