--- src/correlator/basis/dlr_fermi/dlr10000_1e-4.dat
@@ -1,33 +1,0 @@
-           1  -9994.6473080965079      -0.36682460884623214            -3531
-           2  -7663.5078230753579      -0.23946685076890908            -1836
-           3  -5005.3526919034912      -0.13144545365420698             -747
-           4  -3348.2006683710479       -8.7653427436996001E-002        -404
-           5  -2222.2313956372514       -3.9964239215328108E-002        -235
-           6  -1172.4499398371804       -1.9097107554534358E-002        -123
-           7  -641.58433453277939       -1.2974372911340475E-002         -64
-           8  -351.27503008140980       -8.7818757520352446E-003         -35
-           9  -199.82119607664052       -4.1076704266939683E-003         -19
-          10  -129.74372911340475       -2.7391696074061250E-003         -14
-          11  -82.153408533879372       -1.7361182778416026E-003         -10
-          12  -49.955299019160130       -1.0977344690044056E-003          -6
-          13  -29.935577433888117       -6.5394544304122042E-004          -4
-          14  -11.935692221583974       -2.4440198690935013E-004          -3
-          15  -3.9002567407423441       -6.8079993869750845E-005          -2
-          16  -1.7119810046288277       -2.6136190935016125E-007          -1
-          17  -3.2391278384334754E-002   2.6136190935016125E-007           0
-          18   3.7419471792360146        6.8079993869750845E-005           1
-          19   11.428975512206694        2.4179507817422246E-004           2
-          20   18.708347716321022        6.5394544304122042E-004           3
-          21   29.935577433888117        1.0977344690044056E-003           5
-          22   52.315635443297630        1.7361182778416026E-003           9
-          23   80.198041816597438        3.3616100490419937E-003          17
-          24   129.74372911340475        4.7742768886335895E-003          23
-          25   190.97107554534358        6.7232200980839874E-003          32
-          26   328.61363413551749        1.2974372911340475E-002          61
-          27   616.70783273361030        2.6892880392335949E-002         112
-          28   1111.1156978186257        4.1852508354638107E-002         209
-          29   2151.4304313868761        6.7766574615545458E-002         376
-          30   3197.1391372262487       0.11724499398371804              712
-          31   4933.6626618688824       0.19158769557688393             1052
-          32   7663.5078230753579       0.30555372087254973             1896
-          33   9994.6473080965079       0.46897997593487217             4206
