--- src/correlator/basis/dlr_fermi/dlr10000_1e-6.dat
@@ -1,47 +1,0 @@
-           1  -9994.6473080965079      -0.38317539115376786            -8239
-           2  -9148.3645377501725      -0.29258177311249139            -4064
-           3  -7663.5078230753579      -0.22870911344375430            -1815
-           4  -6111.0744174509946      -0.13553314923109092            -1098
-           5  -5005.3526919034912      -0.10379498329072380             -751
-           6  -3348.2006683710479       -7.9928478430656216E-002        -525
-           7  -2222.2313956372514       -5.1897491645361900E-002        -399
-           8  -1674.1003341855239       -3.3883287307772729E-002        -230
-           9  -1172.4499398371804       -2.3948461947110491E-002        -166
-          10  -876.53427436995980       -1.6941643653886364E-002        -122
-          11  -641.58433453277939       -1.1463269026444754E-002         -90
-          12  -351.27503008140980       -7.7088479091701285E-003         -65
-          13  -268.92880392335945       -5.2315635443297634E-003         -48
-          14  -199.82119607664052       -3.7416695432642044E-003         -35
-          15  -129.74372911340475       -2.6157817721648817E-003         -25
-          16  -82.153408533879372       -1.6808050245209968E-003         -18
-          17  -64.871864556702377       -1.0977344690044056E-003         -13
-          18  -49.955299019160130       -6.5394544304122042E-004         -10
-          19  -29.935577433888117       -4.0544915347938984E-004          -6
-          20  -18.708347716321022       -2.4440198690935013E-004          -5
-          21  -11.935692221583974       -1.3005409292360146E-004          -4
-          22  -3.9002567407423441       -6.8079993869750845E-005          -3
-          23  -1.7119810046288277       -2.6136190935016125E-007          -2
-          24  -3.2391278384334754E-002   2.6136190935016125E-007          -1
-          25   3.7419471792360146        3.0292992251101407E-005           0
-          26   6.8479240185153119        8.2832096520610185E-005           1
-          27   11.428975512206694        2.6471318209197444E-004           2
-          28   18.708347716321022        4.4669748719483262E-004           3
-          29   29.935577433888117        7.1645431415279714E-004           4
-          30   40.099020908298719        1.2488824754790034E-003           5
-          31   52.315635443297630        2.1177054567357956E-003           9
-          32   80.198041816597438        2.7391696074061250E-003          12
-          33   129.74372911340475        3.9104317905496021E-003          17
-          34   190.97107554534358        5.2315635443297634E-003          24
-          35   328.61363413551749        7.8083182094503979E-003          33
-          36   458.53076105779013        1.0956678429624500E-002          46
-          37   616.70783273361030        1.6430681706775873E-002          63
-          38   837.05016709276208        2.2926538052889509E-002         117
-          39   1111.1156978186257        3.3883287307772729E-002         157
-          40   1527.7686043627486        4.9923286281502006E-002         214
-          41   2151.4304313868761        7.6388430218137432E-002         289
-          42   3197.1391372262487       0.11111156978186257              388
-          43   4933.6626618688824       0.15277686043627486              517
-          44   6111.0744174509946       0.21514304313868760              738
-          45   7663.5078230753579       0.28102002406512783             1163
-          46   9148.3645377501725       0.36682460884623214             2056
-          47   9994.6473080965079       0.44444627912745027             4223
