--- src/correlator/basis/dlr_fermi/dlr10000_1e-8.dat
@@ -1,59 +1,0 @@
-           1  -9994.6473080965079      -0.39938629025201605            -8372
-           2  -9148.3645377501725      -0.30555372087254973            -4104
-           3  -7663.5078230753579      -0.22222313956372514            -1835
-           4  -6111.0744174509946      -0.16741003341855243            -1415
-           5  -5005.3526919034912      -0.12513381729758727            -1120
-           6  -4151.7993316289521       -9.5793847788441966E-002        -778
-           7  -3348.2006683710479       -7.6388430218137432E-002        -543
-           8  -2810.2002406512784       -5.8622496991859022E-002        -411
-           9  -2222.2313956372514       -4.3826713718498000E-002        -306
-          10  -1674.1003341855239       -2.8588639180469288E-002        -237
-          11  -1172.4499398371804       -2.2926538052889509E-002        -169
-          12  -876.53427436995980       -1.9097107554534358E-002        -124
-          13  -641.58433453277939       -1.3446440196167975E-002         -90
-          14  -478.96923894220987       -9.5485537772671790E-003         -77
-          15  -351.27503008140980       -6.7232200980839874E-003         -65
-          16  -268.92880392335945       -4.7742768886335895E-003         -48
-          17  -199.82119607664052       -3.9437787492124406E-003         -35
-          18  -160.39608363319485       -3.2435932278351187E-003         -26
-          19  -129.74372911340475       -2.2857951024413390E-003         -19
-          20  -82.153408533879372       -1.5601026962969377E-003         -16
-          21  -64.871864556702377       -1.2488824754790034E-003         -14
-          22  -49.955299019160130       -1.0024755227074681E-003         -10
-          23  -38.874856253937800       -6.2444123773950169E-004          -8
-          24  -29.935577433888117       -3.9002567407423443E-004          -7
-          25  -18.708347716321022       -2.3766236932313303E-004          -6
-          26  -11.935692221583974       -6.8079993869750845E-005          -5
-          27  -7.8005134814846881       -2.0572557091974445E-005          -4
-          28  -3.9002567407423441       -2.6136190935016125E-007          -3
-          29  -1.7119810046288277        2.6136190935016125E-007          -2
-          30  -3.2391278384334754E-002   2.0572557091974445E-005          -1
-          31   3.7419471792360146        6.8079993869750845E-005           0
-          32   6.8479240185153119        2.3766236932313303E-004           1
-          33   11.428975512206694        3.9002567407423443E-004           2
-          34   18.708347716321022        6.2444123773950169E-004           3
-          35   23.871384443167948        1.0024755227074681E-003           4
-          36   29.935577433888117        1.5601026962969377E-003           5
-          37   40.099020908298719        2.0538352133469841E-003           6
-          38   52.315635443297630        2.8658172566111886E-003           7
-          39   80.198041816597438        3.9437787492124406E-003           9
-          40   99.910598038320273        5.4783392148122501E-003          13
-          41   129.74372911340475        6.4871864556702375E-003          18
-          42   157.75114996849763        7.6110795733060317E-003          21
-          43   190.97107554534358        1.0463127088659527E-002          25
-          44   249.61643140751005        1.5222159146612063E-002          34
-          45   328.61363413551749        2.1913356859249000E-002          46
-          46   458.53076105779013        2.6892880392335949E-002          63
-          47   616.70783273361030        3.2079216726638979E-002          88
-          48   837.05016709276208        4.7896923894220983E-002         122
-          49   1111.1156978186257        5.7177278360938576E-002         166
-          50   1527.7686043627486        7.0255006016281957E-002         237
-          51   2151.4304313868761        8.7653427436996001E-002         304
-          52   2566.3373381311180       0.11111156978186257              404
-          53   3197.1391372262487       0.15277686043627486              533
-          54   3831.7539115376790       0.20758996658144760              736
-          55   4933.6626618688824       0.24668313309344411              899
-          56   6111.0744174509946       0.29258177311249139             1165
-          57   7663.5078230753579       0.38317539115376786             1507
-          58   9148.3645377501725       0.44444627912745027             2052
-          59   9994.6473080965079       0.49336626618688822             4205
