--- src/correlator/basis/dlr_fermi/dlr1000_1e-10.dat
@@ -1,51 +1,0 @@
-           1  -999.46473080965086      -0.43028608627737519             -758
-           2  -937.95995186974437      -0.36682460884623214             -371
-           3  -860.57217255475030      -0.31971391372262487             -232
-           4  -798.77258050403202      -0.27106629846218183             -164
-           5  -669.64013367420966      -0.22870911344375430             -126
-           6  -525.78181461682789      -0.17530685487399200              -99
-           7  -468.97997593487219      -0.13144545365420698              -80
-           8  -415.17993316289517      -0.11111156978186257              -67
-           9  -319.71391372262485       -9.5793847788441966E-002         -49
-          10  -243.55454634579303       -7.0255006016281957E-002         -41
-          11  -207.58996658144758       -5.1897491645361900E-002         -35
-          12  -175.30685487399199       -3.9964239215328108E-002         -26
-          13  -128.31686690655590       -3.3883287307772729E-002         -22
-          14  -95.793847788441980       -2.9311248495929511E-002         -19
-          15  -79.928478430656213       -2.0926254177319054E-002         -14
-          16  -67.766574615545466       -1.6430681706775873E-002         -12
-          17  -51.897491645361896       -1.1974230973555246E-002         -10
-          18  -39.964239215328107       -9.1431804097653561E-003          -9
-          19  -29.311248495929512       -6.7232200980839874E-003          -8
-          20  -21.913356859248999       -3.5770890865284089E-003          -7
-          21  -15.775114996849762       -2.5809364556702375E-003          -6
-          22  -11.463269026444753       -1.5720892148122496E-003          -5
-          23  -7.8875574984248811       -4.8468787601762251E-004          -4
-          24  -3.2435932278351185       -2.0142042669396825E-004          -3
-          25  -1.5191115556832051       -4.1817905496025799E-006          -2
-          26  -5.1826045414935740E-002   4.1817905496025799E-006          -1
-          27   2.9935577433888119        4.8468787601762251E-004           0
-          28   6.9444731113664107        1.5720892148122496E-003           1
-          29   9.9910598038320266        2.3341607851877508E-003           2
-          30   13.888946222732821        3.4215621239823777E-003           3
-          31   19.097107554534361        5.9871154867776229E-003           4
-          32   26.892880392335947        7.8875574984248811E-003           5
-          33   36.572721639061427        1.0463127088659527E-002           6
-          34   49.923286281502001        1.3888946222732821E-002           7
-          35   58.622496991859023        1.8286360819530712E-002           8
-          36   70.255006016281953        2.2926538052889509E-002           9
-          37   83.705016709276208        2.9311248495929511E-002          12
-          38   99.846572563004003        3.6572721639061424E-002          15
-          39   131.44545365420697        4.7896923894220983E-002          18
-          40   175.30685487399199        6.1670783273361028E-002          25
-          41   207.58996658144758        7.0255006016281957E-002          29
-          42   239.46685076890907        7.9928478430656216E-002          34
-          43   319.71391372262485        9.9846572563004013E-002          48
-          44   399.38629025201601       0.12831686690655592               64
-          45   513.26746762622361       0.16741003341855243               79
-          46   585.16354622498284       0.21514304313868760              103
-          47   669.64013367420966       0.26289090730841397              135
-          48   798.77258050403202       0.31971391372262487              184
-          49   888.89255825490056       0.36682460884623214              250
-          50   937.95995186974437       0.43028608627737519              386
-          51   999.46473080965086       0.49759816005040380              713
