--- src/correlator/basis/dlr_fermi/dlr1000_1e-12.dat
@@ -1,61 +1,0 @@
-           1  -999.46473080965086      -0.46897997593487217            -1000
-           2  -974.21818538317211      -0.43028608627737519             -751
-           3  -937.95995186974437      -0.35061370974798400             -371
-           4  -860.57217255475030      -0.31971391372262487             -289
-           5  -798.77258050403202      -0.28102002406512783             -232
-           6  -669.64013367420966      -0.22222313956372514             -165
-           7  -611.10744174509944      -0.17530685487399200             -126
-           8  -525.78181461682789      -0.15277686043627486              -99
-           9  -468.97997593487219      -0.13553314923109092              -81
-          10  -415.17993316289517      -0.10757152156934380              -68
-          11  -319.71391372262485       -8.3705016709276214E-002         -57
-          12  -281.02002406512781       -7.3145443278122849E-002         -49
-          13  -243.55454634579303       -6.5722726827103492E-002         -41
-          14  -207.58996658144758       -5.1897491645361900E-002         -35
-          15  -175.30685487399199       -4.3826713718498000E-002         -30
-          16  -146.29088655624571       -3.8194215109068716E-002         -26
-          17  -128.31686690655590       -2.8588639180469288E-002         -22
-          18  -95.793847788441980       -2.3948461947110491E-002         -19
-          19  -79.928478430656213       -2.0926254177319054E-002         -16
-          20  -67.766574615545466       -1.5222159146612063E-002         -14
-          21  -51.897491645361896       -1.1463269026444754E-002         -12
-          22  -39.964239215328107       -8.4708218269431822E-003         -10
-          23  -33.883287307772733       -6.2404107851877508E-003          -9
-          24  -29.311248495929512       -4.9955299019160135E-003          -8
-          25  -21.913356859248999       -3.5770890865284089E-003          -7
-          26  -15.775114996849762       -2.5809364556702375E-003          -6
-          27  -11.463269026444753       -1.5720892148122496E-003          -5
-          28  -7.8875574984248811       -4.8468787601762251E-004          -4
-          29  -3.2435932278351185       -2.0142042669396825E-004          -3
-          30  -1.5191115556832051       -4.1817905496025799E-006          -2
-          31  -5.1826045414935740E-002   4.1817905496025799E-006          -1
-          32   2.9935577433888119        2.0142042669396825E-004           0
-          33   6.9444731113664107        4.8468787601762251E-004           1
-          34   9.9910598038320266        1.5720892148122496E-003           2
-          35   13.888946222732821        3.2409097951173220E-003           3
-          36   19.097107554534361        4.3909378760176223E-003           4
-          37   26.892880392335947        5.9871154867776229E-003           5
-          38   31.550229993699524        8.0198041816597447E-003           6
-          39   36.572721639061427        9.9910598038320270E-003           7
-          40   49.923286281502001        1.2974372911340475E-002           8
-          41   58.622496991859023        1.6941643653886364E-002           9
-          42   70.255006016281953        2.1913356859249000E-002          11
-          43   83.705016709276208        2.6892880392335949E-002          13
-          44   99.846572563004003        3.1550229993699525E-002          16
-          45   131.44545365420697        3.6572721639061424E-002          19
-          46   152.77686043627489        4.7896923894220983E-002          23
-          47   175.30685487399199        5.5555784890931284E-002          26
-          48   207.58996658144758        6.4158433453277958E-002          30
-          49   239.46685076890907        7.3145443278122849E-002          35
-          50   281.02002406512781        8.3705016709276214E-002          41
-          51   319.71391372262485        9.9846572563004013E-002          49
-          52   399.38629025201601       0.11724499398371804               57
-          53   457.41822688750858       0.13553314923109092               66
-          54   513.26746762622361       0.15985695686131243               82
-          55   585.16354622498284       0.19158769557688393              108
-          56   669.64013367420966       0.22870911344375430              139
-          57   798.77258050403202       0.26289090730841397              161
-          58   888.89255825490056       0.30555372087254973              189
-          59   937.95995186974437       0.35061370974798400              257
-          60   974.21818538317211       0.41517993316289520              397
-          61   999.46473080965086       0.48710909269158603              759
