--- src/correlator/basis/dlr_fermi/dlr1000_1e-4.dat
@@ -1,24 +1,0 @@
-           1  -999.46473080965086      -0.35061370974798400             -182
-           2  -798.77258050403202      -0.22222313956372514             -102
-           3  -525.78181461682789       -7.6388430218137432E-002         -62
-           4  -319.71391372262485       -4.3826713718498000E-002         -34
-           5  -175.30685487399199       -2.4961643140751003E-002         -24
-           6  -95.793847788441980       -1.3888946222732821E-002         -18
-           7  -51.897491645361896       -7.7088479091701285E-003         -10
-           8  -29.311248495929512       -3.8025979091701285E-003          -6
-           9  -15.775114996849762       -1.5720892148122496E-003          -4
-          10  -7.8875574984248811       -4.8468787601762251E-004          -3
-          11  -3.2435932278351185       -4.1817905496025799E-006          -2
-          12  -5.1826045414935740E-002   4.1817905496025799E-006          -1
-          13   2.9935577433888119        4.8468787601762251E-004           0
-          14   6.9444731113664107        1.8253845132223767E-003           1
-          15   13.888946222732821        3.9437787492124406E-003           2
-          16   26.892880392335947        7.7749712507875594E-003           3
-          17   49.923286281502001        1.5222159146612063E-002           5
-          18   70.255006016281953        2.4961643140751003E-002           9
-          19   99.846572563004003        4.3826713718498000E-002          17
-          20   175.30685487399199        7.3145443278122849E-002          33
-          21   319.71391372262485       0.12177727317289651               64
-          22   513.26746762622361       0.19158769557688393               95
-          23   798.77258050403202       0.30555372087254973              162
-          24   999.46473080965086       0.47893370153781817              372
