--- src/correlator/basis/dlr_fermi/dlr1000_1e-6.dat
@@ -1,35 +1,0 @@
-           1  -999.46473080965086      -0.45741822688750861             -385
-           2  -798.77258050403202      -0.29258177311249139             -184
-           3  -669.64013367420966      -0.22870911344375430             -103
-           4  -525.78181461682789      -0.17530685487399200              -64
-           5  -319.71391372262485      -0.10379498329072380              -48
-           6  -243.55454634579303       -7.6388430218137432E-002         -34
-           7  -175.30685487399199       -5.5555784890931284E-002         -26
-           8  -128.31686690655590       -4.1852508354638107E-002         -19
-           9  -95.793847788441980       -3.1233272837801591E-002         -13
-          10  -67.766574615545466       -2.1913356859249000E-002         -10
-          11  -51.897491645361896       -1.6430681706775873E-002          -8
-          12  -29.311248495929512       -8.7818757520352446E-003          -6
-          13  -21.913356859248999       -6.2404107851877508E-003          -5
-          14  -15.775114996849762       -4.2354109134715911E-003          -4
-          15  -7.8875574984248811       -1.8253845132223767E-003          -3
-          16  -3.2435932278351185       -4.8468787601762251E-004          -2
-          17  -1.5191115556832051       -4.1817905496025799E-006          -1
-          18  -5.1826045414935740E-002   4.1817905496025799E-006           0
-          19   2.9935577433888119        4.8468787601762251E-004           1
-          20   6.9444731113664107        1.5720892148122496E-003           2
-          21   13.888946222732821        3.5770890865284089E-003           3
-          22   19.097107554534361        4.9955299019160135E-003           4
-          23   26.892880392335947        6.7232200980839874E-003           5
-          24   49.923286281502001        1.1974230973555246E-002           7
-          25   70.255006016281953        1.6039608363319489E-002           9
-          26   99.846572563004003        2.0926254177319054E-002          13
-          27   131.44545365420697        3.5127503008140978E-002          18
-          28   175.30685487399199        4.5853076105779017E-002          25
-          29   239.46685076890907        5.8622496991859022E-002          33
-          30   319.71391372262485        9.1706152211558034E-002          47
-          31   513.26746762622361       0.11724499398371804               65
-          32   669.64013367420966       0.14629088655624570               96
-          33   798.77258050403202       0.22222313956372514              161
-          34   937.95995186974437       0.36682460884623214              369
-          35   999.46473080965086       0.45741822688750861              761
