--- src/correlator/basis/dlr_fermi/dlr1000_1e-8.dat
@@ -1,43 +1,0 @@
-           1  -999.46473080965086      -0.47893370153781817             -708
-           2  -937.95995186974437      -0.36682460884623214             -382
-           3  -798.77258050403202      -0.31971391372262487             -183
-           4  -669.64013367420966      -0.27106629846218183             -134
-           5  -525.78181461682789      -0.19969314512600803             -103
-           6  -415.17993316289517      -0.15277686043627486              -79
-           7  -319.71391372262485      -0.12831686690655592              -64
-           8  -243.55454634579303      -0.11111156978186257              -48
-           9  -207.58996658144758       -8.3705016709276214E-002         -34
-          10  -175.30685487399199       -6.1670783273361028E-002         -26
-          11  -128.31686690655590       -4.5853076105779017E-002         -19
-          12  -95.793847788441980       -3.3883287307772729E-002         -16
-          13  -67.766574615545466       -2.5948745822680950E-002         -13
-          14  -51.897491645361896       -1.8286360819530712E-002         -10
-          15  -39.964239215328107       -1.3446440196167975E-002          -8
-          16  -29.311248495929512       -9.9910598038320270E-003          -7
-          17  -21.913356859248999       -7.3278121239823777E-003          -6
-          18  -15.775114996849762       -4.2354109134715911E-003          -5
-          19  -7.8875574984248811       -2.8169700980839869E-003          -4
-          20  -3.2435932278351185       -1.8253845132223767E-003          -3
-          21  -1.5191115556832051       -6.6534020488267804E-004          -2
-          22  -5.1826045414935740E-002  -4.1817905496025799E-006          -1
-          23   2.9935577433888119        4.1817905496025799E-006           0
-          24   6.9444731113664107        4.8468787601762251E-004           1
-          25   9.9910598038320266        1.5720892148122496E-003           2
-          26   13.888946222732821        3.5770890865284089E-003           3
-          27   19.097107554534361        4.9955299019160135E-003           4
-          28   26.892880392335947        7.1471597951173220E-003           5
-          29   36.572721639061427        1.1974230973555246E-002           6
-          30   49.923286281502001        1.6039608363319489E-002           7
-          31   70.255006016281953        2.0926254177319054E-002           9
-          32   83.705016709276208        2.7777892445465642E-002          13
-          33   99.846572563004003        3.3883287307772729E-002          15
-          34   131.44545365420697        5.5555784890931284E-002          18
-          35   175.30685487399199        7.0255006016281957E-002          25
-          36   239.46685076890907        8.7653427436996001E-002          33
-          37   319.71391372262485       0.11435455672187715               47
-          38   399.38629025201601       0.14051001203256391               65
-          39   513.26746762622361       0.17530685487399200               97
-          40   669.64013367420966       0.21514304313868760              124
-          41   798.77258050403202       0.28102002406512783              162
-          42   937.95995186974437       0.35061370974798400              369
-          43   999.46473080965086       0.43028608627737519              755
