--- src/correlator/basis/dlr_fermi/dlr100_1e-10.dat
@@ -1,31 +1,0 @@
-           1  -99.946473080965092      -0.43028608627737519              -71
-           2  -95.786740307563633      -0.38317539115376786              -36
-           3  -88.889255825490054      -0.33482006683710486              -24
-           4  -79.877258050403213      -0.26289090730841397              -17
-           5  -73.364921769246422      -0.20758996658144760              -13
-           6  -58.516354622498277      -0.15985695686131243              -10
-           7  -45.741822688750858      -0.13553314923109092               -9
-           8  -38.317539115376789      -0.11435455672187715               -8
-           9  -30.555372087254973       -7.9928478430656216E-002          -7
-          10  -23.448998796743609       -5.5555784890931284E-002          -6
-          11  -18.341230442311605       -3.5127503008140978E-002          -5
-          12  -11.435455672187715       -2.0647491645361900E-002          -4
-          13  -7.9928478430656211       -1.0602508354638104E-002          -3
-          14  -5.5555784890931283       -3.8775030081409801E-003          -2
-          15  -1.7563751504070488       -3.3454324396820639E-005          -1
-          16  0.43571196076640539        3.3454324396820639E-005           0
-          17   2.8588639180469286        3.8775030081409801E-003           1
-          18   5.7177278360938573        1.0602508354638104E-002           2
-          19   9.1706152211558027        2.2535760784671895E-002           3
-          20   13.144545365420699        3.5127503008140978E-002           4
-          21   17.530685487399197        6.0888636586448254E-002           5
-          22   22.870911344375429        7.6388430218137432E-002           6
-          23   29.258177311249138        9.9846572563004013E-002           7
-          24   41.517993316289520       0.12831686690655592                8
-          25   47.893370153781817       0.15985695686131243               10
-          26   56.204004813025570       0.19969314512600803               12
-          27   66.964013367420961       0.23946685076890908               15
-          28   79.877258050403213       0.30555372087254973               22
-          29   91.483645377501716       0.38317539115376786               28
-          30   97.421818538317211       0.43028608627737519               39
-          31   99.946473080965092       0.47893370153781817               74
