--- src/correlator/basis/dlr_fermi/dlr100_1e-12.dat
@@ -1,35 +1,0 @@
-           1  -99.946473080965092      -0.47893370153781817              -73
-           2  -95.786740307563633      -0.43028608627737519              -40
-           3  -88.889255825490054      -0.38317539115376786              -29
-           4  -79.877258050403213      -0.30555372087254973              -23
-           5  -73.364921769246422      -0.24355454634579302              -16
-           6  -58.516354622498277      -0.19969314512600803              -13
-           7  -52.578181461682796      -0.15985695686131243              -11
-           8  -45.741822688750858      -0.12831686690655592              -10
-           9  -38.317539115376789       -9.9846572563004013E-002          -9
-          10  -30.555372087254973       -7.6388430218137432E-002          -8
-          11  -23.448998796743609       -5.9866712692227271E-002          -7
-          12  -18.341230442311605       -3.3883287307772729E-002          -6
-          13  -14.629088655624569       -2.2535760784671895E-002          -5
-          14  -11.435455672187715       -1.0602508354638104E-002          -4
-          15  -7.9928478430656211       -3.8775030081409801E-003          -3
-          16  -5.5555784890931283       -1.6113634135517460E-003          -2
-          17  -1.7563751504070488       -3.3454324396820639E-005          -1
-          18  0.43571196076640539        3.3454324396820639E-005           0
-          19   2.8588639180469286        1.6113634135517460E-003           1
-          20   5.7177278360938573        3.8775030081409801E-003           2
-          21   9.1706152211558027        1.0602508354638104E-002           3
-          22   13.144545365420699        2.0647491645361900E-002           4
-          23   17.530685487399197        3.5127503008140978E-002           5
-          24   22.870911344375429        5.1897491645361900E-002           6
-          25   29.258177311249138        6.5722726827103492E-002           7
-          26   35.061370974798393        7.9928478430656216E-002           8
-          27   41.517993316289520       0.11111156978186257                9
-          28   47.893370153781817       0.15277686043627486               10
-          29   56.204004813025570       0.17530685487399200               12
-          30   66.964013367420961       0.19969314512600803               16
-          31   79.877258050403213       0.25663373381311183               19
-          32   86.057217255475024       0.30555372087254973               23
-          33   91.483645377501716       0.35061370974798400               35
-          34   97.421818538317211       0.43028608627737519               70
-          35   99.946473080965092       0.46897997593487217              100
