--- src/correlator/basis/dlr_fermi/dlr100_1e-4.dat
@@ -1,16 +1,0 @@
-           1  -99.946473080965092      -0.39938629025201605              -35
-           2  -73.364921769246422      -0.27106629846218183              -16
-           3  -45.741822688750858      -0.18341230442311607               -9
-           4  -30.555372087254973       -9.1706152211558034E-002          -6
-           5  -18.341230442311605       -4.3826713718498000E-002          -4
-           6  -11.435455672187715       -1.2576713718497997E-002          -3
-           7  -5.5555784890931283       -3.8775030081409801E-003          -2
-           8  -1.7563751504070488       -3.3454324396820639E-005          -1
-           9  0.43571196076640539        3.3454324396820639E-005           0
-          10   2.8588639180469286        3.8775030081409801E-003           1
-          11   9.1706152211558027        1.2576713718497997E-002           2
-          12   17.530685487399197        3.9964239215328108E-002           3
-          13   29.258177311249138        7.3145443278122849E-002           5
-          14   56.204004813025570       0.12177727317289651                8
-          15   79.877258050403213       0.29258177311249139               15
-          16   99.946473080965092       0.44444627912745027               30
