--- src/correlator/basis/dlr_fermi/dlr100_1e-6.dat
@@ -1,21 +1,0 @@
-           1  -99.946473080965092      -0.47893370153781817              -40
-           2  -88.889255825490054      -0.31971391372262487              -24
-           3  -73.364921769246422      -0.25663373381311183              -16
-           4  -58.516354622498277      -0.19969314512600803              -12
-           5  -45.741822688750858      -0.12620091997479810               -9
-           6  -30.555372087254973       -7.3145443278122849E-002          -6
-           7  -18.341230442311605       -3.8194215109068716E-002          -5
-           8  -11.435455672187715       -2.4305784890931284E-002          -4
-           9  -5.5555784890931283       -1.0602508354638104E-002          -3
-          10  -1.7563751504070488       -3.8775030081409801E-003          -2
-          11  0.43571196076640539       -3.3454324396820639E-005          -1
-          12   2.8588639180469286        3.3454324396820639E-005           0
-          13   5.7177278360938573        3.8775030081409801E-003           1
-          14   9.1706152211558027        1.0602508354638104E-002           2
-          15   17.530685487399197        2.2535760784671895E-002           3
-          16   29.258177311249138        3.9964239215328108E-002           4
-          17   41.517993316289520        8.3705016709276214E-002           5
-          18   56.204004813025570       0.11724499398371804                9
-          19   79.877258050403213       0.15985695686131243               12
-          20   91.483645377501716       0.24355454634579302               16
-          21   99.946473080965092       0.36682460884623214               35
