--- src/correlator/basis/dlr_fermi/dlr100_1e-8.dat
@@ -1,26 +1,0 @@
-           1  -99.946473080965092      -0.41517993316289520              -75
-           2  -95.786740307563633      -0.33482006683710486              -40
-           3  -88.889255825490054      -0.27106629846218183              -23
-           4  -73.364921769246422      -0.22222313956372514              -16
-           5  -58.516354622498277      -0.17530685487399200              -11
-           6  -45.741822688750858      -0.14051001203256391               -9
-           7  -30.555372087254973      -0.10757152156934380               -7
-           8  -23.448998796743609       -6.4158433453277958E-002          -6
-           9  -18.341230442311605       -3.5127503008140978E-002          -5
-          10  -11.435455672187715       -2.2535760784671895E-002          -4
-          11  -7.9928478430656211       -1.0602508354638104E-002          -3
-          12  -5.5555784890931283       -3.8775030081409801E-003          -2
-          13  -1.7563751504070488       -3.3454324396820639E-005          -1
-          14  0.43571196076640539        3.3454324396820639E-005           0
-          15   2.8588639180469286        3.8775030081409801E-003           1
-          16   5.7177278360938573        1.0602508354638104E-002           2
-          17   9.1706152211558027        2.0647491645361900E-002           3
-          18   17.530685487399197        3.8194215109068716E-002           4
-          19   22.870911344375429        5.9866712692227271E-002           5
-          20   29.258177311249138        8.3705016709276214E-002           7
-          21   41.517993316289520       0.11724499398371804                9
-          22   56.204004813025570       0.15985695686131243               12
-          23   66.964013367420961       0.25663373381311183               16
-          24   79.877258050403213       0.31971391372262487               23
-          25   91.483645377501716       0.41517993316289520               35
-          26   99.946473080965092       0.49973236540482546               70
